fix: filebeat with empty API
