+++ src/handler/http/router/mod.rs
@@ -155,13 +155,13 @@ pub fn get_service_routes(cfg: &mut web::ServiceConfig) {
             .service(organization::org_summary)
             .service(organization::get_user_passcode)
             .service(organization::update_user_passcode)
-            .service(organization::org_es_index)
-            .service(organization::org_es_license)
-            .service(organization::org_es_xpack)
-            .service(organization::org_es_index_template)
-            .service(organization::org_es_index_template_create)
-            .service(organization::org_es_data_stream)
-            .service(organization::org_es_data_stream_create)
+            .service(organization::es::org_index)
+            .service(organization::es::org_license)
+            .service(organization::es::org_xpack)
+            .service(organization::es::org_index_template)
+            .service(organization::es::org_index_template_create)
+            .service(organization::es::org_data_stream)
+            .service(organization::es::org_data_stream_create)
             .service(users::update)
             .service(templates::save_template)
             .service(templates::get_template)
